Move select and select modal to separate components
