+++ app/scripts/react-directives/network-connections/networkConnectionsPage.jsx
@@ -4,7 +4,7 @@ import { observer } from 'mobx-react-lite';
 import { useTranslation, Trans } from 'react-i18next';
 import ConnectionsEditor from './connectionsEditor';
 import SwitcherEditor from './switcherEditor';
-import { SelectModal } from './modals';
+import { SelectModal } from '../components/modals/selectModal';
 import ConfirmModal from '../components/modals/confirmModal';
 import { HorizontalTabs, TabContent, TabItem, TabPane, TabsList } from './tabs';
